[refactor] always pass AssetItems onAssetSelected
